EES-4908 Move PublicationController.ListPublicationDataSets to new DataSetsController.ListDataSets
