+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Requests/Public.Data/DataSetListRequest.cs
@@ -1,18 +1,15 @@
 #nullable enable
+using System;
 using FluentValidation;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Requests.Public.Data;
 
 public record DataSetListRequest
 {
-    /// <summary>
-    /// The page of results to fetch.
-    /// </summary>
+    public Guid PublicationId { get; init; }
+
     public int Page { get; init; } = 1;
 
-    /// <summary>
-    /// The maximum number of results per page.
-    /// </summary>
     public int PageSize { get; init; } = 10;
 
     public class Validator : AbstractValidator<DataSetListRequest>
